Update state appropriately when unit does not exist
Summary:
If a unit that existed in the past no longer exists, we must inform Terraform
of that fact so that it can plan to re-create it.

Ref T39.

Test Plan:
Deployed the test phabricator deployment I'm working on. Deleted the units with
fleetctl behind Terraform's back. Ran `terraform plan`, and saw that it was
planning to re-create the units I deleted.

Reviewers: kian, #developers

Reviewed By: kian, #developers

Maniphest Tasks: T39

Differential Revision: https://phabricator.paperg.com/D30