Consolidate everything into one package
Summary:
This layout is a more normal and has the benefit of working with
CircleCI without any explicit configuration.

The GitHub project has been renamed to terraform-provider-fleet (was
terraform-fleet), which makes the canonical go import path and the
binary name that Terraform expects for the plugin to work all aligned.

Also fix a little error turned up by `go vet`. Yay tests!

Test Plan:
`go build`. It fails, but in the same way it did before. T346 will fix
that.

Reviewers: #developers, kian

Reviewed By: #developers, kian

Differential Revision: https://phabricator.paperg.com/D244