Update to be compatible with recent fleet
Summary:
This makes the provider compatible with the current version of fleet.
The HTTP API is the default now and the etcd API was moved around in the
fleet package so this provider would no longer build. Now it does.

Resolves T346.

Test Plan:
I had a CoreOS cluster already set up, and I made a test Terraform
project with the example in the README. Applied it, and verified that the unit
was created with `fleetctl`:

    $ terraform apply
    fleet_unit.myapp: Creating...
    [...]
    Apply complete! Resources: 1 added, 0 changed, 0 destroyed.
    $ fleetctl --tunnel 54.85.85.76 list-units
    UNIT            MACHINE                 ACTIVE  SUB
    myapp.service   355bc6c2.../10.0.9.2    active  running

Reviewers: #developers, kian

Reviewed By: #developers, kian

Maniphest Tasks: T346

Differential Revision: https://phabricator.paperg.com/D245